Moved lookups params to types file
